Split reth code into submodules
