+++ crates/utils/sov-rpc-eth-types/src/lib.rs
@@ -1,771 +1,10 @@
-use alloy_eips::BlockId;
-use alloy_primitives::{Address, Bytes, B256, U256};
-use alloy_rpc_types::error::EthRpcErrorCode;
-use alloy_rpc_types::request::TransactionInputError;
-use alloy_sol_types::{ContractError, RevertReason};
-use alloy_transport::{RpcError, TransportErrorKind};
-use core::time::Duration;
-use reth_errors::RethError;
-use reth_primitives_traits::transaction::error::InvalidTransactionError;
-use reth_primitives_traits::transaction::signed::RecoveryError;
-use revm::context::result::{HaltReason, InvalidTransaction, OutOfGasError};
-use revm::context_interface::result::{EVMError, InvalidHeader};
-use std::convert::Infallible;
+mod eth_api_error;
+mod revert_error;
+mod rpc_invalid_transaction_error;
+mod to_rpc_error;
+mod utils;
 
-/// An error due to invalid transaction.
-///
-/// The only reason this exists is to maintain compatibility with other clients de-facto standard
-/// error messages.
-///
-/// These error variants can be thrown when the transaction is checked prior to execution.
-///
-/// These variants also cover all errors that can be thrown by revm.
-///
-/// ## Nomenclature
-///
-/// This type is explicitly modeled after geth's error variants and uses
-///   `fee cap` for `max_fee_per_gas`
-///   `tip` for `max_priority_fee_per_gas`
-#[derive(thiserror::Error, Debug)]
-pub enum RpcInvalidTransactionError {
-    /// returned if the nonce of a transaction is lower than the one present in the local chain.
-    #[error("nonce too low: next nonce {state}, tx nonce {tx}")]
-    NonceTooLow {
-        /// The nonce of the transaction.
-        tx: u64,
-        /// The current state of the nonce in the local chain.
-        state: u64,
-    },
-    /// returned if the nonce of a transaction is higher than the next one expected based on the
-    /// local chain.
-    #[error("nonce too high")]
-    NonceTooHigh,
-    /// Returned if the nonce of a transaction is too high
-    /// Incrementing the nonce would lead to invalid state (overflow)
-    #[error("nonce has max value")]
-    NonceMaxValue,
-    /// thrown if the transaction sender doesn't have enough funds for a transfer
-    #[error("insufficient funds for transfer")]
-    InsufficientFundsForTransfer,
-    /// thrown if creation transaction provides the init code bigger than init code size limit.
-    #[error("max initcode size exceeded")]
-    MaxInitCodeSizeExceeded,
-    /// Represents the inability to cover max fee + value (account balance too low).
-    #[error("insufficient funds for gas * price + value: have {balance} want {cost}")]
-    InsufficientFunds {
-        /// Transaction cost.
-        cost: U256,
-        /// Current balance of transaction sender.
-        balance: U256,
-    },
-    /// This is similar to [`Self::InsufficientFunds`] but with a different error message and
-    /// exists for compatibility reasons.
-    ///
-    /// This error is used in `eth_estimateCall` when the highest available gas limit, capped with
-    /// the allowance of the caller is too low: [`Self::GasTooLow`].
-    #[error("gas required exceeds allowance ({gas_limit})")]
-    GasRequiredExceedsAllowance {
-        /// The gas limit the transaction was executed with.
-        gas_limit: u64,
-    },
-    /// Thrown when calculating gas usage
-    #[error("gas uint64 overflow")]
-    GasUintOverflow,
-    /// Thrown if the transaction is specified to use less gas than required to start the
-    /// invocation.
-    #[error("intrinsic gas too low")]
-    GasTooLow,
-    /// Thrown if the transaction gas exceeds the limit
-    #[error("intrinsic gas too high")]
-    GasTooHigh,
-    /// Thrown if the transaction gas limit exceeds the maximum
-    #[error("gas limit too high")]
-    GasLimitTooHigh,
-    /// Thrown if a transaction is not supported in the current network configuration.
-    #[error("transaction type not supported")]
-    TxTypeNotSupported,
-    /// Thrown to ensure no one is able to specify a transaction with a tip higher than the total
-    /// fee cap.
-    #[error("max priority fee per gas higher than max fee per gas")]
-    TipAboveFeeCap,
-    /// A sanity error to avoid huge numbers specified in the tip field.
-    #[error("max priority fee per gas higher than 2^256-1")]
-    TipVeryHigh,
-    /// A sanity error to avoid huge numbers specified in the fee cap field.
-    #[error("max fee per gas higher than 2^256-1")]
-    FeeCapVeryHigh,
-    /// Thrown post London if the transaction's fee is less than the base fee of the block
-    #[error("max fee per gas less than block base fee")]
-    FeeCapTooLow,
-    /// Thrown if the sender of a transaction is a contract.
-    #[error("sender is not an EOA")]
-    SenderNoEOA,
-    /// Gas limit was exceeded during execution.
-    /// Contains the gas limit.
-    #[error("out of gas: gas required exceeds: {0}")]
-    BasicOutOfGas(u64),
-    /// Gas limit was exceeded during memory expansion.
-    /// Contains the gas limit.
-    #[error("out of gas: gas exhausted during memory expansion: {0}")]
-    MemoryOutOfGas(u64),
-    /// Gas limit was exceeded during precompile execution.
-    /// Contains the gas limit.
-    #[error("out of gas: gas exhausted during precompiled contract execution: {0}")]
-    PrecompileOutOfGas(u64),
-    /// An operand to an opcode was invalid or out of range.
-    /// Contains the gas limit.
-    #[error("out of gas: invalid operand to an opcode: {0}")]
-    InvalidOperandOutOfGas(u64),
-    /// Thrown if executing a transaction failed during estimate/call
-    #[error(transparent)]
-    Revert(RevertError),
-    /// Unspecific EVM halt error.
-    #[error("EVM error: {0:?}")]
-    EvmHalt(HaltReason),
-    /// Invalid chain id set for the transaction.
-    #[error("invalid chain ID")]
-    InvalidChainId,
-    /// The transaction is before Spurious Dragon and has a chain ID
-    #[error("transactions before Spurious Dragon should not have a chain ID")]
-    OldLegacyChainId,
-    /// The transitions is before Berlin and has access list
-    #[error("transactions before Berlin should not have access list")]
-    AccessListNotSupported,
-    /// `max_fee_per_blob_gas` is not supported for blocks before the Cancun hardfork.
-    #[error("max_fee_per_blob_gas is not supported for blocks before the Cancun hardfork")]
-    MaxFeePerBlobGasNotSupported,
-    /// `blob_hashes`/`blob_versioned_hashes` is not supported for blocks before the Cancun
-    /// hardfork.
-    #[error("blob_versioned_hashes is not supported for blocks before the Cancun hardfork")]
-    BlobVersionedHashesNotSupported,
-    /// Block `blob_base_fee` is greater than tx-specified `max_fee_per_blob_gas` after Cancun.
-    #[error("max fee per blob gas less than block blob gas fee")]
-    BlobFeeCapTooLow,
-    /// Blob transaction has a versioned hash with an invalid blob
-    #[error("blob hash version mismatch")]
-    BlobHashVersionMismatch,
-    /// Blob transaction has no versioned hashes
-    #[error("blob transaction missing blob hashes")]
-    BlobTransactionMissingBlobHashes,
-    /// Blob transaction has too many blobs
-    #[error("blob transaction exceeds max blobs per block; got {have}")]
-    TooManyBlobs {
-        /// The number of blobs in the transaction.
-        have: usize,
-    },
-    /// Blob transaction is a create transaction
-    #[error("blob transaction is a create transaction")]
-    BlobTransactionIsCreate,
-    /// EIP-7702 is not enabled.
-    #[error("EIP-7702 authorization list not supported")]
-    AuthorizationListNotSupported,
-    /// EIP-7702 transaction has invalid fields set.
-    #[error("EIP-7702 authorization list has invalid fields")]
-    AuthorizationListInvalidFields,
-    /// Transaction priority fee is below the minimum required priority fee.
-    #[error("transaction priority fee below minimum required priority fee {minimum_priority_fee}")]
-    PriorityFeeBelowMinimum {
-        /// Minimum required priority fee.
-        minimum_priority_fee: u128,
-    },
-    /// Any other error
-    #[error("{0}")]
-    Other(Box<dyn ToRpcError>),
-}
-
-impl RpcInvalidTransactionError {
-    /// crates a new [`RpcInvalidTransactionError::Other`] variant.
-    pub fn other<E: ToRpcError>(err: E) -> Self {
-        Self::Other(Box::new(err))
-    }
-
-    /// Returns the rpc error code for this error.
-    pub const fn error_code(&self) -> i32 {
-        match self {
-            Self::InvalidChainId
-            | Self::GasTooLow
-            | Self::GasTooHigh
-            | Self::GasRequiredExceedsAllowance { .. }
-            | Self::NonceTooLow { .. }
-            | Self::NonceTooHigh { .. }
-            | Self::FeeCapTooLow
-            | Self::FeeCapVeryHigh => EthRpcErrorCode::InvalidInput.code(),
-            Self::Revert(_) => EthRpcErrorCode::ExecutionError.code(),
-            _ => EthRpcErrorCode::TransactionRejected.code(),
-        }
-    }
-
-    /// Converts the halt error
-    ///
-    /// Takes the configured gas limit of the transaction which is attached to the error
-    pub const fn halt(reason: HaltReason, gas_limit: u64) -> Self {
-        match reason {
-            HaltReason::OutOfGas(err) => Self::out_of_gas(err, gas_limit),
-            HaltReason::NonceOverflow => Self::NonceMaxValue,
-            err => Self::EvmHalt(err),
-        }
-    }
-
-    /// Converts the out of gas error
-    pub const fn out_of_gas(reason: OutOfGasError, gas_limit: u64) -> Self {
-        match reason {
-            OutOfGasError::Basic | OutOfGasError::ReentrancySentry => {
-                Self::BasicOutOfGas(gas_limit)
-            }
-            OutOfGasError::Memory | OutOfGasError::MemoryLimit => Self::MemoryOutOfGas(gas_limit),
-            OutOfGasError::Precompile => Self::PrecompileOutOfGas(gas_limit),
-            OutOfGasError::InvalidOperand => Self::InvalidOperandOutOfGas(gas_limit),
-        }
-    }
-
-    /// Converts this error into the rpc error object.
-    pub fn into_rpc_err(self) -> jsonrpsee_types::error::ErrorObject<'static> {
-        self.into()
-    }
-}
-
-impl From<RpcInvalidTransactionError> for jsonrpsee_types::error::ErrorObject<'static> {
-    fn from(err: RpcInvalidTransactionError) -> Self {
-        match err {
-            RpcInvalidTransactionError::Revert(revert) => {
-                // include out data if some
-                rpc_err(
-                    revert.error_code(),
-                    revert.to_string(),
-                    revert.output.as_ref().map(|out| out.as_ref()),
-                )
-            }
-            RpcInvalidTransactionError::Other(err) => err.to_rpc_error(),
-            err => rpc_err(err.error_code(), err.to_string(), None),
-        }
-    }
-}
-
-impl From<InvalidTransaction> for RpcInvalidTransactionError {
-    fn from(err: InvalidTransaction) -> Self {
-        match err {
-            InvalidTransaction::InvalidChainId | InvalidTransaction::MissingChainId => {
-                Self::InvalidChainId
-            }
-            InvalidTransaction::PriorityFeeGreaterThanMaxFee => Self::TipAboveFeeCap,
-            InvalidTransaction::GasPriceLessThanBasefee => Self::FeeCapTooLow,
-            InvalidTransaction::CallerGasLimitMoreThanBlock
-            | InvalidTransaction::TxGasLimitGreaterThanCap { .. } => {
-                // tx.gas > block.gas_limit
-                Self::GasTooHigh
-            }
-            InvalidTransaction::CallGasCostMoreThanGasLimit { .. } => {
-                // tx.gas < cost
-                Self::GasTooLow
-            }
-            InvalidTransaction::GasFloorMoreThanGasLimit { .. } => {
-                // Post prague EIP-7623 tx floor calldata gas cost > tx.gas_limit
-                // where floor gas is the minimum amount of gas that will be spent
-                // In other words, the tx's gas limit is lower that the minimum gas requirements of
-                // the tx's calldata
-                Self::GasTooLow
-            }
-            InvalidTransaction::RejectCallerWithCode => Self::SenderNoEOA,
-            InvalidTransaction::LackOfFundForMaxFee { fee, balance } => Self::InsufficientFunds {
-                cost: *fee,
-                balance: *balance,
-            },
-            InvalidTransaction::OverflowPaymentInTransaction => Self::GasUintOverflow,
-            InvalidTransaction::NonceOverflowInTransaction => Self::NonceMaxValue,
-            InvalidTransaction::CreateInitCodeSizeLimit => Self::MaxInitCodeSizeExceeded,
-            InvalidTransaction::NonceTooHigh { .. } => Self::NonceTooHigh,
-            InvalidTransaction::NonceTooLow { tx, state } => Self::NonceTooLow { tx, state },
-            InvalidTransaction::AccessListNotSupported => Self::AccessListNotSupported,
-            InvalidTransaction::MaxFeePerBlobGasNotSupported => Self::MaxFeePerBlobGasNotSupported,
-            InvalidTransaction::BlobVersionedHashesNotSupported => {
-                Self::BlobVersionedHashesNotSupported
-            }
-            InvalidTransaction::BlobGasPriceGreaterThanMax => Self::BlobFeeCapTooLow,
-            InvalidTransaction::EmptyBlobs => Self::BlobTransactionMissingBlobHashes,
-            InvalidTransaction::BlobVersionNotSupported => Self::BlobHashVersionMismatch,
-            InvalidTransaction::TooManyBlobs { have, .. } => Self::TooManyBlobs { have },
-            InvalidTransaction::BlobCreateTransaction => Self::BlobTransactionIsCreate,
-            InvalidTransaction::AuthorizationListNotSupported => {
-                Self::AuthorizationListNotSupported
-            }
-            InvalidTransaction::AuthorizationListInvalidFields
-            | InvalidTransaction::EmptyAuthorizationList => Self::AuthorizationListInvalidFields,
-            InvalidTransaction::Eip2930NotSupported
-            | InvalidTransaction::Eip1559NotSupported
-            | InvalidTransaction::Eip4844NotSupported
-            | InvalidTransaction::Eip7702NotSupported
-            | InvalidTransaction::Eip7873NotSupported => Self::TxTypeNotSupported,
-            InvalidTransaction::Eip7873MissingTarget => {
-                Self::other(internal_rpc_err(err.to_string()))
-            }
-        }
-    }
-}
-
-impl From<InvalidTransactionError> for RpcInvalidTransactionError {
-    fn from(err: InvalidTransactionError) -> Self {
-        use InvalidTransactionError;
-        // This conversion is used to convert any transaction errors that could occur inside the
-        // txpool (e.g. `eth_sendRawTransaction`) to their corresponding RPC
-        match err {
-            InvalidTransactionError::InsufficientFunds(res) => Self::InsufficientFunds {
-                cost: res.expected,
-                balance: res.got,
-            },
-            InvalidTransactionError::NonceNotConsistent { tx, state } => {
-                Self::NonceTooLow { tx, state }
-            }
-            InvalidTransactionError::OldLegacyChainId => {
-                // Note: this should be unreachable since Spurious Dragon now enabled
-                Self::OldLegacyChainId
-            }
-            InvalidTransactionError::ChainIdMismatch => Self::InvalidChainId,
-            InvalidTransactionError::Eip2930Disabled
-            | InvalidTransactionError::Eip1559Disabled
-            | InvalidTransactionError::Eip4844Disabled
-            | InvalidTransactionError::Eip7702Disabled
-            | InvalidTransactionError::TxTypeNotSupported => Self::TxTypeNotSupported,
-            InvalidTransactionError::GasUintOverflow => Self::GasUintOverflow,
-            InvalidTransactionError::GasTooLow => Self::GasTooLow,
-            InvalidTransactionError::GasTooHigh => Self::GasTooHigh,
-            InvalidTransactionError::TipAboveFeeCap => Self::TipAboveFeeCap,
-            InvalidTransactionError::FeeCapTooLow => Self::FeeCapTooLow,
-            InvalidTransactionError::SignerAccountHasBytecode => Self::SenderNoEOA,
-            InvalidTransactionError::GasLimitTooHigh => Self::GasLimitTooHigh,
-        }
-    }
-}
-
-/// Represents a reverted transaction and its output data.
-///
-/// Displays "execution reverted(: reason)?" if the reason is a string.
-#[derive(Debug, Clone, thiserror::Error)]
-pub struct RevertError {
-    /// The transaction output data
-    ///
-    /// Note: this is `None` if output was empty
-    output: Option<Bytes>,
-}
-
-// === impl RevertError ==
-
-impl RevertError {
-    /// Wraps the output bytes
-    ///
-    /// Note: this is intended to wrap an revm output
-    pub fn new(output: Bytes) -> Self {
-        if output.is_empty() {
-            Self { output: None }
-        } else {
-            Self {
-                output: Some(output),
-            }
-        }
-    }
-
-    /// Returns error code to return for this error.
-    pub const fn error_code(&self) -> i32 {
-        EthRpcErrorCode::ExecutionError.code()
-    }
-}
-
-impl std::fmt::Display for RevertError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str("execution reverted")?;
-        if let Some(reason) = self
-            .output
-            .as_ref()
-            .and_then(|out| RevertReason::decode(out))
-        {
-            let error = reason.to_string();
-            let mut error = error.as_str();
-            if matches!(
-                reason,
-                RevertReason::ContractError(ContractError::Revert(_))
-            ) {
-                // we strip redundant `revert: ` prefix from the revert reason
-                error = error.trim_start_matches("revert: ");
-            }
-            write!(f, ": {error}")?;
-        }
-        Ok(())
-    }
-}
-
-/// Constructs a JSON-RPC error, consisting of `code`, `message` and optional `data`.
-pub fn rpc_err(
-    code: i32,
-    msg: impl Into<String>,
-    data: Option<&[u8]>,
-) -> jsonrpsee_types::error::ErrorObject<'static> {
-    jsonrpsee_types::error::ErrorObject::owned(
-        code,
-        msg.into(),
-        data.map(|data| {
-            jsonrpsee_core::to_json_raw_value(&alloy_primitives::hex::encode_prefixed(data))
-                .expect("serializing String can't fail")
-        }),
-    )
-}
-
-/// Constructs an internal JSON-RPC error.
-pub fn internal_rpc_err(msg: impl Into<String>) -> jsonrpsee_types::error::ErrorObject<'static> {
-    rpc_err(jsonrpsee_types::error::INTERNAL_ERROR_CODE, msg, None)
-}
-
-/// A trait to convert an error to an RPC error.
-pub trait ToRpcError: core::error::Error + Send + Sync + 'static {
-    /// Converts the error to a JSON-RPC error object.
-    fn to_rpc_error(&self) -> jsonrpsee_types::ErrorObject<'static>;
-}
-
-impl ToRpcError for jsonrpsee_types::ErrorObject<'static> {
-    fn to_rpc_error(&self) -> jsonrpsee_types::ErrorObject<'static> {
-        self.clone()
-    }
-}
-
-impl ToRpcError for RpcError<TransportErrorKind> {
-    fn to_rpc_error(&self) -> jsonrpsee_types::ErrorObject<'static> {
-        match self {
-            Self::ErrorResp(payload) => jsonrpsee_types::error::ErrorObject::owned(
-                payload.code as i32,
-                payload.message.clone(),
-                payload.data.clone(),
-            ),
-            err => internal_rpc_err(err.to_string()),
-        }
-    }
-}
-
-/// Result alias
-pub type EthResult<T> = Result<T, EthApiError>;
-
-/// Errors that can occur when interacting with the `eth_` namespace
-#[derive(Debug, thiserror::Error)]
-pub enum EthApiError {
-    /// When a raw transaction is empty
-    #[error("empty transaction data")]
-    EmptyRawTransactionData,
-    /// When decoding a signed transaction fails
-    #[error("failed to decode signed transaction")]
-    FailedToDecodeSignedTransaction,
-    /// When the transaction signature is invalid
-    #[error("invalid transaction signature")]
-    InvalidTransactionSignature,
-    // /// Errors related to the transaction pool
-    // #[error(transparent)]
-    // PoolError(RpcPoolError),
-    /// Header not found for block hash/number/tag
-    #[error("header not found")]
-    HeaderNotFound(BlockId),
-    /// Header range not found for start block hash/number/tag to end block hash/number/tag
-    #[error("header range not found, start block {0:?}, end block {1:?}")]
-    HeaderRangeNotFound(BlockId, BlockId),
-    /// Thrown when historical data is not available because it has been pruned
-    ///
-    /// This error is intended for use as a standard response when historical data is
-    /// requested that has been pruned according to the node's data retention policy.
-    ///
-    /// See also <https://eips.ethereum.org/EIPS/eip-4444>
-    #[error("pruned history unavailable")]
-    PrunedHistoryUnavailable,
-    /// Receipts not found for block hash/number/tag
-    #[error("receipts not found")]
-    ReceiptsNotFound(BlockId),
-    /// Thrown when an unknown block or transaction index is encountered
-    #[error("unknown block or tx index")]
-    UnknownBlockOrTxIndex,
-    /// When an invalid block range is provided
-    #[error("invalid block range")]
-    InvalidBlockRange,
-    /// Thrown when the target block for proof computation exceeds the maximum configured window.
-    #[error("distance to target block exceeds maximum proof window")]
-    ExceedsMaxProofWindow,
-    /// An internal error where prevrandao is not set in the evm's environment
-    #[error("prevrandao not in the EVM's environment after merge")]
-    PrevrandaoNotSet,
-    /// `excess_blob_gas` is not set for Cancun and above
-    #[error("excess blob gas missing in the EVM's environment after Cancun")]
-    ExcessBlobGasNotSet,
-    /// Thrown when a call or transaction request (`eth_call`, `eth_estimateGas`,
-    /// `eth_sendTransaction`) contains conflicting fields (legacy, EIP-1559)
-    #[error("both gasPrice and (maxFeePerGas or maxPriorityFeePerGas) specified")]
-    ConflictingFeeFieldsInRequest,
-    /// Errors related to invalid transactions
-    #[error(transparent)]
-    InvalidTransaction(#[from] RpcInvalidTransactionError),
-    // /// Thrown when constructing an RPC block from primitive block data fails
-    // #[error(transparent)]
-    // InvalidBlockData(#[from] BlockError),
-    /// Thrown when an `AccountOverride` contains conflicting `state` and `stateDiff` fields
-    #[error("account {0:?} has both 'state' and 'stateDiff'")]
-    BothStateAndStateDiffInOverride(Address),
-    /// Other internal error
-    #[error(transparent)]
-    Internal(RethError),
-    // /// Error related to signing
-    // #[error(transparent)]
-    // Signing(#[from] SignError),
-    /// Thrown when a requested transaction is not found
-    #[error("transaction not found")]
-    TransactionNotFound,
-    /// Some feature is unsupported
-    #[error("unsupported")]
-    Unsupported(&'static str),
-    /// General purpose error for invalid params
-    #[error("{0}")]
-    InvalidParams(String),
-    /// When the tracer config does not match the tracer
-    #[error("invalid tracer config")]
-    InvalidTracerConfig,
-    /// When the percentile array is invalid
-    #[error("invalid reward percentiles")]
-    InvalidRewardPercentiles,
-    /// Error thrown when a spawned blocking task failed to deliver an anticipated response.
-    ///
-    /// This only happens if the blocking task panics and is aborted before it can return a
-    /// response back to the request handler.
-    #[error("internal blocking task error")]
-    InternalBlockingTaskError,
-    /// Error thrown when a spawned blocking task failed to deliver an anticipated response
-    #[error("internal eth error")]
-    InternalEthError,
-    /// Error thrown when a (tracing) call exceeds the configured timeout
-    #[error("execution aborted (timeout = {0:?})")]
-    ExecutionTimedOut(Duration),
-    /// Internal Error thrown by the javascript tracer
-    #[error("{0}")]
-    InternalJsTracerError(String),
-    #[error(transparent)]
-    /// Call Input error when both `data` and `input` fields are set and not equal.
-    TransactionInputError(#[from] TransactionInputError),
-    /// Evm generic purpose error.
-    #[error("Revm error: {0}")]
-    EvmCustom(String),
-    /// Bytecode override is invalid.
-    ///
-    /// This can happen if bytecode provided in an
-    /// [`AccountOverride`](alloy_rpc_types_eth::state::AccountOverride) is malformed, e.g. invalid
-    /// 7702 bytecode.
-    // #[error("Invalid bytecode: {0}")]
-    // InvalidBytecode(String),
-    /// Error encountered when converting a transaction type
-    #[error("Transaction conversion error")]
-    TransactionConversionError,
-    // /// Error thrown when tracing with a muxTracer fails
-    // #[error(transparent)]
-    // MuxTracerError(#[from] MuxError),
-    /// Error thrown when waiting for transaction confirmation times out
-    #[error(
-        "Transaction {hash} was added to the mempool but wasn't confirmed within {duration:?}."
-    )]
-    TransactionConfirmationTimeout {
-        /// Hash of the transaction that timed out
-        hash: B256,
-        /// Duration that was waited before timing out
-        duration: Duration,
-    },
-    // /// Error thrown when batch tx response channel fails
-    // #[error(transparent)]
-    // BatchTxRecvError(#[from] RecvError),
-    /// Error thrown when batch tx send channel fails
-    #[error("Batch transaction sender channel closed")]
-    BatchTxSendError,
-    /// Any other error
-    #[error("{0}")]
-    Other(Box<dyn ToRpcError>),
-}
-
-impl EthApiError {
-    /// crates a new [`EthApiError::Other`] variant.
-    pub fn other<E: ToRpcError>(err: E) -> Self {
-        Self::Other(Box::new(err))
-    }
-
-    /// Returns `true` if error is [`RpcInvalidTransactionError::GasTooHigh`]
-    pub const fn is_gas_too_high(&self) -> bool {
-        matches!(
-            self,
-            Self::InvalidTransaction(
-                RpcInvalidTransactionError::GasTooHigh
-                    | RpcInvalidTransactionError::GasLimitTooHigh
-            )
-        )
-    }
-
-    /// Returns `true` if error is [`RpcInvalidTransactionError::GasTooLow`]
-    pub const fn is_gas_too_low(&self) -> bool {
-        matches!(
-            self,
-            Self::InvalidTransaction(RpcInvalidTransactionError::GasTooLow)
-        )
-    }
-
-    /// Returns the [`RpcInvalidTransactionError`] if this is a [`EthApiError::InvalidTransaction`]
-    pub const fn as_invalid_transaction(&self) -> Option<&RpcInvalidTransactionError> {
-        match self {
-            Self::InvalidTransaction(e) => Some(e),
-            _ => None,
-        }
-    }
-
-    /// Converts this error into the rpc error object.
-    pub fn into_rpc_err(self) -> jsonrpsee_types::error::ErrorObject<'static> {
-        self.into()
-    }
-}
-
-impl From<EthApiError> for jsonrpsee_types::error::ErrorObject<'static> {
-    fn from(error: EthApiError) -> Self {
-        match error {
-            EthApiError::FailedToDecodeSignedTransaction
-            | EthApiError::InvalidTransactionSignature
-            | EthApiError::EmptyRawTransactionData
-            | EthApiError::InvalidBlockRange
-            | EthApiError::ExceedsMaxProofWindow
-            | EthApiError::ConflictingFeeFieldsInRequest
-            | EthApiError::BothStateAndStateDiffInOverride(_)
-            | EthApiError::InvalidTracerConfig
-            | EthApiError::TransactionConversionError
-            | EthApiError::InvalidRewardPercentiles => invalid_params_rpc_err(error.to_string()),
-            EthApiError::InvalidTransaction(err) => err.into(),
-            EthApiError::PrevrandaoNotSet
-            | EthApiError::ExcessBlobGasNotSet
-            | EthApiError::Internal(_)
-            | EthApiError::EvmCustom(_) => internal_rpc_err(error.to_string()),
-            EthApiError::UnknownBlockOrTxIndex | EthApiError::TransactionNotFound => {
-                rpc_error_with_code(EthRpcErrorCode::ResourceNotFound.code(), error.to_string())
-            }
-            // TODO(onbjerg): We rewrite the error message here because op-node does string matching
-            // on the error message.
-            //
-            // Until https://github.com/ethereum-optimism/optimism/pull/11759 is released, this must be kept around.
-            EthApiError::HeaderNotFound(id) => rpc_error_with_code(
-                EthRpcErrorCode::ResourceNotFound.code(),
-                format!("block not found: {}", block_id_to_str(id)),
-            ),
-            EthApiError::ReceiptsNotFound(id) => rpc_error_with_code(
-                EthRpcErrorCode::ResourceNotFound.code(),
-                format!("{error}: {}", block_id_to_str(id)),
-            ),
-            EthApiError::HeaderRangeNotFound(start_id, end_id) => rpc_error_with_code(
-                EthRpcErrorCode::ResourceNotFound.code(),
-                format!(
-                    "{error}: start block: {}, end block: {}",
-                    block_id_to_str(start_id),
-                    block_id_to_str(end_id),
-                ),
-            ),
-            err @ EthApiError::TransactionConfirmationTimeout { .. } => rpc_error_with_code(
-                EthRpcErrorCode::TransactionConfirmationTimeout.code(),
-                err.to_string(),
-            ),
-            EthApiError::Unsupported(msg) => internal_rpc_err(msg),
-            EthApiError::InternalJsTracerError(msg) => internal_rpc_err(msg),
-            EthApiError::InvalidParams(msg) => invalid_params_rpc_err(msg),
-            err @ EthApiError::ExecutionTimedOut(_) => rpc_error_with_code(
-                jsonrpsee_types::error::CALL_EXECUTION_FAILED_CODE,
-                err.to_string(),
-            ),
-            err @ (EthApiError::InternalBlockingTaskError | EthApiError::InternalEthError) => {
-                internal_rpc_err(err.to_string())
-            }
-            err @ EthApiError::TransactionInputError(_) => invalid_params_rpc_err(err.to_string()),
-            EthApiError::PrunedHistoryUnavailable => rpc_error_with_code(4444, error.to_string()),
-            EthApiError::Other(err) => err.to_rpc_error(),
-            EthApiError::BatchTxSendError => {
-                internal_rpc_err("Batch transaction sender channel closed".to_string())
-            }
-        }
-    }
-}
-
-#[cfg(feature = "js-tracer")]
-impl From<revm_inspectors::tracing::js::JsInspectorError> for EthApiError {
-    fn from(error: revm_inspectors::tracing::js::JsInspectorError) -> Self {
-        match error {
-            err @ revm_inspectors::tracing::js::JsInspectorError::JsError(_) => {
-                Self::InternalJsTracerError(err.to_string())
-            }
-            err => Self::InvalidParams(err.to_string()),
-        }
-    }
-}
-
-impl From<InvalidHeader> for EthApiError {
-    fn from(value: InvalidHeader) -> Self {
-        match value {
-            InvalidHeader::ExcessBlobGasNotSet => Self::ExcessBlobGasNotSet,
-            InvalidHeader::PrevrandaoNotSet => Self::PrevrandaoNotSet,
-        }
-    }
-}
-
-impl<T> From<EVMError<T, InvalidTransaction>> for EthApiError
-where
-    T: Into<Self>,
-{
-    fn from(err: EVMError<T, InvalidTransaction>) -> Self {
-        match err {
-            EVMError::Transaction(invalid_tx) => match invalid_tx {
-                InvalidTransaction::NonceTooLow { tx, state } => {
-                    Self::InvalidTransaction(RpcInvalidTransactionError::NonceTooLow { tx, state })
-                }
-                _ => RpcInvalidTransactionError::from(invalid_tx).into(),
-            },
-            EVMError::Header(err) => err.into(),
-            EVMError::Database(err) => err.into(),
-            EVMError::Custom(err) => Self::EvmCustom(err),
-        }
-    }
-}
-
-impl From<RecoveryError> for EthApiError {
-    fn from(_: RecoveryError) -> Self {
-        Self::InvalidTransactionSignature
-    }
-}
-
-impl From<Infallible> for EthApiError {
-    fn from(_: Infallible) -> Self {
-        unreachable!()
-    }
-}
-
-/// Constructs an invalid params JSON-RPC error.
-pub fn invalid_params_rpc_err(
-    msg: impl Into<String>,
-) -> jsonrpsee_types::error::ErrorObject<'static> {
-    rpc_err(jsonrpsee_types::error::INVALID_PARAMS_CODE, msg, None)
-}
-
-/// Constructs an internal JSON-RPC error with data
-pub fn internal_rpc_err_with_data(
-    msg: impl Into<String>,
-    data: &[u8],
-) -> jsonrpsee_types::error::ErrorObject<'static> {
-    rpc_err(jsonrpsee_types::error::INTERNAL_ERROR_CODE, msg, Some(data))
-}
-
-/// Constructs an internal JSON-RPC error with code and message
-pub fn rpc_error_with_code(
-    code: i32,
-    msg: impl Into<String>,
-) -> jsonrpsee_types::error::ErrorObject<'static> {
-    rpc_err(code, msg, None)
-}
-
-/// Formats a [`BlockId`] into an error message.
-pub fn block_id_to_str(id: BlockId) -> String {
-    match id {
-        BlockId::Hash(h) => {
-            if h.require_canonical == Some(true) {
-                format!("canonical hash {}", h.block_hash)
-            } else {
-                format!("hash {}", h.block_hash)
-            }
-        }
-        BlockId::Number(n) => format!("{n}"),
-    }
-}
+pub use eth_api_error::{EthApiError, EthResult};
+pub use revert_error::RevertError;
+pub use rpc_invalid_transaction_error::RpcInvalidTransactionError;
+pub use to_rpc_error::ToRpcError;
