+++ crates/utils/sov-rpc-eth-types/src/revert_error.rs
@@ -1,0 +1,59 @@
+use alloy_primitives::Bytes;
+use alloy_rpc_types::error::EthRpcErrorCode;
+use alloy_sol_types::{ContractError, RevertReason};
+
+/// Represents a reverted transaction and its output data.
+///
+/// Displays "execution reverted(: reason)?" if the reason is a string.
+#[derive(Debug, Clone, thiserror::Error)]
+pub struct RevertError {
+    /// The transaction output data
+    ///
+    /// Note: this is `None` if output was empty
+    pub(crate) output: Option<Bytes>,
+}
+
+// === impl RevertError ==
+
+impl RevertError {
+    /// Wraps the output bytes
+    ///
+    /// Note: this is intended to wrap an revm output
+    pub fn new(output: Bytes) -> Self {
+        if output.is_empty() {
+            Self { output: None }
+        } else {
+            Self {
+                output: Some(output),
+            }
+        }
+    }
+
+    /// Returns error code to return for this error.
+    pub const fn error_code(&self) -> i32 {
+        EthRpcErrorCode::ExecutionError.code()
+    }
+}
+
+impl std::fmt::Display for RevertError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_str("execution reverted")?;
+        if let Some(reason) = self
+            .output
+            .as_ref()
+            .and_then(|out| RevertReason::decode(out))
+        {
+            let error = reason.to_string();
+            let mut error = error.as_str();
+            if matches!(
+                reason,
+                RevertReason::ContractError(ContractError::Revert(_))
+            ) {
+                // we strip redundant `revert: ` prefix from the revert reason
+                error = error.trim_start_matches("revert: ");
+            }
+            write!(f, ": {error}")?;
+        }
+        Ok(())
+    }
+}
